mybook/requirements.txt
django==4.2.6
django-ckeditor==6.7.0
Pillow==10.1.0
djangorestframework==3.14.0
requests==2.31.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

channels==4.0.0
channels-redis==4.2.0

Django==4.2.6
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-jazzmin==3.0.0
django-js-asset==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docker==7.0.0

google-generativeai==0.5.2

Jinja2==3.1.2

mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.0


Pillow==10.1.0

redis==5.0.1
requests==2.31.0

template==0.7.6

# pip install -r requirements.txt
# pip list